# fhir-iq-prior-auth/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FHIR IQ Prior Authorization Support API
  description: |
    FHIR-based Prior Authorization Support (PAS) API implementing Da Vinci PAS and DTR
    Implementation Guides for CMS-0057-F compliance demonstration.

    This API provides endpoints for:
    - Submitting prior authorization requests (PAS)
    - Inquiring about authorization status
    - Documentation Templates and Rules (DTR) support
    - Provider and Patient Access APIs
    - Payer-to-payer bulk export operations
  version: 1.0.0-poc
  contact:
    name: FHIR IQ Integration Team
    email: integration@fhir-iq.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://pas-api.fhir-iq.com/fhir
    description: Production FHIR IQ PAS Server
  - url: http://localhost:3000/fhir
    description: Local development server
security:
  - OAuth2:
      - system/Claim.write
      - user/Claim.read
      - patient/Claim.read
paths:
...
  /QuestionnaireResponse:
    post:
      summary: Submit Questionnaire Response
      description: Submit completed DTR questionnaire response
      operationId: createQuestionnaireResponse
      security:
        - OAuth2:
            - user/QuestionnaireResponse.write
            - system/QuestionnaireResponse.write
      requestBody:
        required: true
        content:
          application/fhir+json:
            schema:
              $ref: '#/components/schemas/QuestionnaireResponse'
      responses:
        '201':
          description: QuestionnaireResponse created
          content:
            application/fhir+json:
              schema:
                $ref: '#/components/schemas/QuestionnaireResponse'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.fhir-iq.com/oauth2/authorize
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.read: Read claim resources (system)
            system/Claim.write: Write claim resources (system)
            user/Claim.read: Read claim resources (user)
            user/ClaimResponse.read: Read claim response resources (user)
            patient/Claim.read: Read patient's claim resources
            patient/ClaimResponse.read: Read patient's claim responses
            user/Questionnaire.read: Read questionnaire resources
            user/QuestionnaireResponse.write: Create questionnaire responses
            system/Patient.read: Read patient resources (system)
            system/export: Perform bulk export operations
        clientCredentials:
          tokenUrl: https://auth.fhir-iq.com/oauth2/token
          scopes:
            system/Claim.write: Write claim resources (system)
            system/export: Perform bulk export operations
  schemas:
...
    QuestionnaireResponse:
      type: object
      properties:
        resourceType:
          type: string
          enum: [QuestionnaireResponse]
        status:
          type: string
          enum: [in-progress, completed, amended, entered-in-error, stopped]
        questionnaire:
          type: string
          format: uri
...